agent/requirements.txt
psutil==5.9.6
requests==2.31.0
pyyaml==6.0.1
pydantic==2.5.0
